@mbasmanova 2023-12-14
@xumingming Thank you for working on this.

> kRelaxedPrefix: _a_bc%%
> kRelaxedSuffix: %%_a_bc
> kRelaxedSubstring: %%_a_bc%%

I haven't read the code yet, but I understand how to handle first 2 patterns, but I'm not sure what's the approach for implementing efficient matching for the last pattern. Would you clarify?