@xumingming 2023-12-15
@mbasmanova Removed kRelaxedSubstring optimization from this PR and updated description.