@mbasmanova 2023-12-20
Seeing lot of linter warnings. Some examples:

<img width="429" alt="Screenshot 2023-12-19 at 7 45 03 PM" src="https://github.com/facebookincubator/velox/assets/27965151/85aa8df1-3e59-48e7-8aa0-a895f8fad009">
<img width="462" alt="Screenshot 2023-12-19 at 7 45 24 PM" src="https://github.com/facebookincubator/velox/assets/27965151/25664757-0696-409f-8b52-5c3d0b581362">
<img width="481" alt="Screenshot 2023-12-19 at 7 45 56 PM" src="https://github.com/facebookincubator/velox/assets/27965151/e816ddaa-c05c-4395-9a50-a95c64e7683e">
<img width="470" alt="Screenshot 2023-12-19 at 7 46 14 PM" src="https://github.com/facebookincubator/velox/assets/27965151/e5555f18-f08a-44d9-8428-1b7df247c245">