@mbasmanova 2023-12-20
FYI, benchmark results I'm seeing

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
like_generic##like_generic                                   1.10s   906.81m
like_prefix##like_prefix                                   12.77ms     78.32
like_prefix##like_relaxed_prefix_1                         16.32ms     61.26
like_prefix##like_relaxed_prefix_2                         13.57ms     73.69
like_prefix##starts_with                                   13.21ms     75.69
like_substring##like_substring                             28.42ms     35.19
like_substring##strpos                                     22.80ms     43.86
like_suffix##like_suffix                                   13.63ms     73.38
like_suffix##like_relaxed_suffix_1                         17.26ms     57.95
like_suffix##like_relaxed_suffix_2                         14.42ms     69.35
like_suffix##ends_with                                     13.31ms     75.13
```