@mbasmanova 2023-12-20
@xumingming James, I'm fixing lint warnings and will try to merge this PR. I have a question, do you expect this to handle queries like

```
<column> LIKE "concat"("concat"('%\u0040', <other column>), '%')
```