@mbasmanova 2023-12-20
Linter shows more warnings:

<img width="460" alt="Screenshot 2023-12-19 at 8 46 32 PM" src="https://github.com/facebookincubator/velox/assets/27965151/4737d808-cb29-40c3-bab3-4347114d16f8">