@xumingming 2023-12-20
> Linter shows more warnings:
> 
> <img alt="Screenshot 2023-12-19 at 8 46 32 PM" width="460" src="https://private-user-images.githubusercontent.com/27965151/291755918-4737d808-cb29-40c3-bab3-4347114d16f8.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTEiLCJleHAiOjE3MDMwMzgwMTQsIm5iZiI6MTcwMzAzNzcxNCwicGF0aCI6Ii8yNzk2NTE1MS8yOTE3NTU5MTgtNDczN2Q4MDgtY2IyOS00MGMzLWJhYjMtNDM0NzExNGQxNmY4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFJV05KWUFYNENTVkVINTNBJTJGMjAyMzEyMjAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjMxMjIwVDAyMDE1NFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTU2NjViMGJlY2ExYzJhZmJmOTVlM2ZiYzgyMTM0ZGFhZGE2ZDA1Y2JiZmYzMTgyODEwNmJiNDBkN2NkODM5MGQmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.xjGSW8ZISbju3S_ZHbg-ECFyiHBrGI9LuuVFjIBOyrU">

@mbasmanova Sorry for the warnings. is plugin publicly available, so I can fix these warnings myself?