@mbasmanova 2023-12-20
> is plugin publicly available

I don't think so.