@mbasmanova 2023-12-20
> You mean there is unicode in the pattern? I suppose no currently, currently we are iterating the pattern as char* which does not handle unicode?

I see. Would you double check that at least we get correct results, even if we don't use optimized code path?