@xumingming 2023-12-20
> > You mean there is unicode in the pattern? I suppose no currently, currently we are iterating the pattern as char* which does not handle unicode?
> 
> I see. Would you double check that at least we get correct results, even if we don't use optimized code path?

Double checked, results are:

* Presto LIKE has bug supporting unicode: https://github.com/prestodb/presto/issues/21577
* Added code to limit current fast path for only ascii pattern.
* Created an issue to optimize unicode pattern in the future: https://github.com/facebookincubator/velox/issues/8125