@xumingming 2023-12-20
> > The speedup for kRelaxedPrefix is about 40x, speedup for kRelaxedSuffix is about 100x.
> 
> @xumingming The original optimization showed speed up of ~20x, but this is one showing much higher speedup. Is there some explanation?

~20x was from `Debug` build, ~100x from `Release` build