@xumingming 2024-01-09
@mbasmanova Updated this PR to handle UTF8, introduced two functions: `matchRelaxedFixedPatternForward` and `matchRelaxedFixedPatternBackward`.

-  `matchRelaxedFixedPatternForward` is used to handle `kRelaxedFixed`, `kRelaxedPrefix` and `kRelaxedSuffix + ascii`.
  - It has different implementations for ascii & utf8.
-  `matchRelaxedFixedPatternBackward` is used to handle `kRelaxedSuffix + utf8`.
  - It only has utf8 implmentation, because ascii input does not need to use this function.