@xumingming 2024-01-24
> @xumingming Thank you for working on this. Looks fine to me % some nits. I feel we'll have to refactor the code for readability if we decide to add more logic.

@mbasmanova Sure, I plan to implement kRelaxedSubstring, I can do some refactoring before that.