@mbasmanova 2024-01-25
> I plan to implement kRelaxedSubstring, I can do some refactoring before that.

@xumingming Sounds good. Thanks.