@xumingming 2024-01-25
Updated the code to resolve all the review comments and rebased main.