@xumingming 2024-01-25
> @xumingming format-check is failing. Would you take a look?
> 
> Curious, would you like to write a blog post about all these optimizations for LIKE?

Fixed the format issue, and glad to write a blog post for LIKE :)