@mbasmanova 2024-01-25
> glad to write a blog post for LIKE :)

Great. Here is an example of a PR that adds a blog post:

https://github.com/facebookincubator/velox/pull/6851

and here is a link to the blog: 

https://velox-lib.io/blog/reduce-agg

Looking forward to a blog post about LIKE optimizations. Thanks.