@mbasmanova 2024-01-25
Seeing errors:

```
fbcode/velox/functions/lib/Re2Functions.cpp:1516:24: runtime error: load of value 272, which is not a valid value for type 'facebook::velox::functions::SubPatternKind'
    #0 0x7f732f7e4ee7 in facebook::velox::functions::parsePattern[abi:cxx11](std::basic_string_view<char, std::char_traits<char>>, std::optional<char>, std::vector<facebook::velox::functions::SubPatternKind, std::allocator<facebook::velox::functions::SubPatternKind>>&, std::vector<std::pair<unsigned long, unsigned long>, std::allocator<std::pair<unsigned long, unsigned long>>>&) fbcode/velox/functions/lib/Re2Functions.cpp:1516

SUMMARY: UndefinedBehaviorSanitizer: invalid-enum-load fbcode/velox/functions/lib/Re2Functions.cpp:1516:24 in 
```